services/edge-gateway/app/app/gateway_legacy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
asyncpg==0.29.0
clickhouse-driver==0.2.6
prometheus-client==0.19.0
python-multipart==0.0.6
httpx==0.26.0
aiokafka==0.10.0
redis[hiredis]==5.0.1
jinja2==3.1.4
avro-python3==1.10.2
minio==7.2.5
pandas==2.2.2
pyarrow==16.1.0
# GraphQL dependencies
graphene==3.3
graphql-core==3.2.3

